requirements.txt
flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
tqdm==4.66.2
msal==1.31.1
